refactor: separate optimizer
